date doesn't work. Update on rest
- Filter the logs
- Separate into different boxes of logs (Done)
- sumamrise with a report
- add colours (added colors but a line has no color)
- EndDayView details
- Summary View? (too late to do)
- Hard coding of footprint (Sch)

+++ LogManager.swift
@@ -10,10 +10,10 @@ import Foundation
 struct CarbonLog: Identifiable, Codable {
     var id = UUID()
     var name: [String]
-    var footprint: [Int]
+    var footprint: [Double]
     var notes: [String]
     var date: Date
-    var totalFootPrint: Int 
+    var totalFootPrint: Double
 }
 
 class CarbonLogManager: ObservableObject {
